# docs/swagger.yaml
basePath: /v1
definitions:
  models.Login:
    properties:
      password:
        type: string
      user_name:
        type: string
    type: object
  models.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
info:
  contact:
    email: debidarmawan1998@gmail.com
  description: This is API Documentation of Majoo Technical Assessment
  license: {}
  title: Majoo Technical Assessment Swagger - Debi Darmawan
paths:
  /login:
    post:
      description: Majoo Assessment User Login
      parameters:
      - description: User Login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Login'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Majoo Assessment User Login
      tags:
      - Users
  /merchant/omzet:
    get:
      description: Majoo Assessment Get Merchant Omzet
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Period
        in: query
        name: period
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Majoo Assessment Get Merchant Omzet
      tags:
      - Merchants
  /merchant/outlet/omzet:
    get:
      description: Majoo Assessment Get Merchant Outlet Omzet
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Outlet ID
        in: query
        name: outlet_id
        type: integer
      - description: Period
        in: query
        name: period
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Majoo Assessment Get Merchant Outlet Omzet
      tags:
      - Merchants
swagger: "2.0"